Move the button layout code to ArrangeOverride
